better code structure
